Seperate system into cpu, memory, and host
