+++ examples/cpu_percent.rs
@@ -1,7 +1,7 @@
 use std::{thread, time};
 
 fn main() {
-    let mut cpu_percent_collector = match psutil::system::CpuPercentCollector::new() {
+    let mut cpu_percent_collector = match psutil::cpu::CpuPercentCollector::new() {
         Ok(cpu_percent_collector) => cpu_percent_collector,
         Err(_) => {
             println!("Could not initialize the CpuPercentCollector object");
