+++ examples/status.rs
@@ -3,5 +3,5 @@
 #[cfg(not(test))]
 fn main() {
     println!("Processes: {}", psutil::process::all().unwrap().len());
-    println!("System uptime: {} seconds", psutil::system::uptime());
+    println!("System uptime: {} seconds", psutil::host::uptime());
 }
