+++ src/lib.rs
@@ -3,11 +3,13 @@
 #[macro_use]
 mod utils;
 
+pub mod cpu;
 pub mod disk;
+pub mod host;
+pub mod memory;
 pub mod network;
 pub mod pidfile;
 pub mod process;
-pub mod system;
 
 /// Type for process identifiers.
 ///
